requirements.txt
google-api-python-client
mysql-connector-python
selenium
